+++ packages/localizations/src/pt-BR.ts
@@ -129,6 +129,7 @@ export const ptBR: LocalizationResource = {
     },
     membersPage: {
       action__invite: 'Convidar',
+      action__search: 'Pesquisar',
       activeMembersTab: {
         menuAction__remove: 'Remover membro',
         tableHeader__actions: 'Ações',
@@ -460,6 +461,7 @@ export const ptBR: LocalizationResource = {
       actionText: 'Não possui uma conta?',
       actionText__join_waitlist: 'Quer ser notificado quando estivermos prontos?',
       subtitle: 'para continuar em {{applicationName}}',
+      subtitleCombined: undefined,
       title: 'Entrar',
       titleCombined: undefined,
     },
@@ -543,7 +545,9 @@ export const ptBR: LocalizationResource = {
       actionLink__use_phone: 'Ou use telefone',
       actionText: 'Possui uma conta?',
       subtitle: 'para continuar em {{applicationName}}',
+      subtitleCombined: 'para continuar em {{applicationName}}',
       title: 'Criar sua conta',
+      titleCombined: 'Criar sua conta',
     },
   },
   socialButtonsBlockButton: 'Continuar com {{provider|titleize}}',
@@ -701,6 +705,11 @@ export const ptBR: LocalizationResource = {
         resendButton: 'Não recebeu um código? Reenviar',
         successMessage: 'O e-mail {{identifier}} foi adicionado na sua conta.',
       },
+      enterpriseSSOLink: {
+        formButton: 'Clique para autenticar',
+        formSubtitle: 'Complete a autenticação com {{identifier}}',
+      },
+      formHint: 'Você precisará verificar este endereço de email antes de poder adicioná-lo à sua conta.',
       removeResource: {
         messageLine1: '{{identifier}} será removido desta conta.',
         messageLine2: 'Você não conseguirá fazer login novamente com este endereço de e-mail.',