+++ packages/localizations/src/ro-RO.ts
@@ -131,6 +131,7 @@ export const roRO: LocalizationResource = {
     },
     membersPage: {
       action__invite: 'Invitați',
+      action__search: undefined,
       activeMembersTab: {
         menuAction__remove: 'Îndepărtați membrul',
         tableHeader__actions: undefined,
@@ -462,6 +463,7 @@ export const roRO: LocalizationResource = {
       actionText: 'Nu aveți cont?',
       actionText__join_waitlist: undefined,
       subtitle: 'pentru a continua la {{applicationName}}',
+      subtitleCombined: undefined,
       title: 'Conectați-vă',
       titleCombined: undefined,
     },
@@ -541,7 +543,9 @@ export const roRO: LocalizationResource = {
       actionLink__use_phone: undefined,
       actionText: 'Aveți un cont?',
       subtitle: 'pentru a continua la {{applicationName}}',
+      subtitleCombined: 'pentru a continua la {{applicationName}}',
       title: 'Creați-vă un cont',
+      titleCombined: 'Creați-vă un cont',
     },
   },
   socialButtonsBlockButton: 'Continuați cu {{provider|titleize}}',
@@ -700,6 +704,11 @@ export const roRO: LocalizationResource = {
         resendButton: 'Nu ați primit un link? Trimiteți din nou',
         successMessage: 'E-mailul {{identifier}} a fost adăugat în contul dvs.',
       },
+      enterpriseSSOLink: {
+        formButton: undefined,
+        formSubtitle: undefined,
+      },
+      formHint: undefined,
       removeResource: {
         messageLine1: '{{identifier}} va fi eliminat din acest cont.',
         messageLine2: 'Nu veți mai putea să vă conectați folosind această adresă de e-mail.',