+++ packages/localizations/src/vi-VN.ts
@@ -129,6 +129,7 @@ export const viVN: LocalizationResource = {
     },
     membersPage: {
       action__invite: 'Mời',
+      action__search: undefined,
       activeMembersTab: {
         menuAction__remove: 'Gỡ bỏ thành viên',
         tableHeader__actions: undefined,
@@ -458,6 +459,7 @@ export const viVN: LocalizationResource = {
       actionText: 'Chưa có tài khoản?',
       actionText__join_waitlist: undefined,
       subtitle: 'để tiếp tục với {{applicationName}}',
+      subtitleCombined: undefined,
       title: 'Đăng nhập',
       titleCombined: undefined,
     },
@@ -537,7 +539,9 @@ export const viVN: LocalizationResource = {
       actionLink__use_phone: undefined,
       actionText: 'Đã có tài khoản?',
       subtitle: 'để tiếp tục với {{applicationName}}',
+      subtitleCombined: 'để tiếp tục với {{applicationName}}',
       title: 'Tạo tài khoản của bạn',
+      titleCombined: 'Tạo tài khoản của bạn',
     },
   },
   socialButtonsBlockButton: 'Tiếp tục với {{provider|titleize}}',
@@ -694,6 +698,11 @@ export const viVN: LocalizationResource = {
         resendButton: 'Không nhận được liên kết? Gửi lại',
         successMessage: 'Email {{identifier}} đã được thêm vào tài khoản của bạn.',
       },
+      enterpriseSSOLink: {
+        formButton: undefined,
+        formSubtitle: undefined,
+      },
+      formHint: undefined,
       removeResource: {
         messageLine1: '{{identifier}} sẽ bị xóa khỏi tài khoản này.',
         messageLine2: 'Bạn sẽ không thể đăng nhập bằng địa chỉ email này nữa.',