+++ packages/nextjs/src/app-router/server/currentUser.ts
@@ -4,6 +4,7 @@ import { clerkClient } from '../../server/clerkClient';
 import { auth } from './auth';
 
 export async function currentUser(): Promise<User | null> {
+  // eslint-disable-next-line @typescript-eslint/no-require-imports
   require('server-only');
 
   const { userId } = await auth();